the transaction keep-alive or its request monitor is now being passed to wherever it is needed

+++ src/migration/DatabaseMigrator.js
@@ -2,6 +2,7 @@
 import ObjectStoreMigrator from "./ObjectStoreMigrator"
 import DatabaseSchema from "../schema/DatabaseSchema"
 import UpgradedDatabaseSchema from "../schema/UpgradedDatabaseSchema"
+import KeepAlive from "../transaction/KeepAlive"
 import ReadOnlyTransaction from "../transaction/ReadOnlyTransaction"
 import Transaction from "../transaction/Transaction"
 
@@ -12,7 +13,9 @@ const FIELDS = Object.freeze({
   database: Symbol("database"),
   transaction: Symbol("transaction"),
   schemaDescriptors: Symbol("schemaDescriptors"),
-  currentVersion: Symbol("currentVersion")
+  currentVersion: Symbol("currentVersion"),
+  keepAliveObjectStore: Symbol("keepAliveObjectStore"),
+  commitDelay: Symbol("commitDelay")
 })
 
 /**
@@ -32,8 +35,11 @@ export default class DatabaseMigrator {
    * @param {number} currentVersion The current version of the database, as a
    *        positive integer, or set to {@code 0} if the database is being
    *        created.
+   * @param {number} commitDelay The delay in milliseconds how long the
+   *        transaction should be kept alive if inactive.
    */
-  constructor(database, transaction, schemaDescriptors, currentVersion) {
+  constructor(database, transaction, schemaDescriptors, currentVersion,
+      commitDelay) {
     if (!schemaDescriptors.length) {
       throw new Error("The list of schema descriptors cannot be empty")
     }
@@ -83,6 +89,25 @@ export default class DatabaseMigrator {
      * @type {number}
      */
     this[FIELDS.currentVersion] = currentVersion
+    
+    let keepAliveObjectStore
+    keepAliveObjectStore = generateKeepAliveObjectStoreName(sortedSchemasCopy)
+    
+    /**
+     * The name of the object store to use to keep the {@code versionchange}
+     * transaction alive as long as needed.
+     * 
+     * @type {string}
+     */
+    this[FIELDS.keepAliveObjectStore] = keepAliveObjectStore
+    
+    /**
+     * Delay in milliseconds for how long an inactive transction should be
+     * kept alive.
+     * 
+     * @type {number}
+     */
+    this[FIELDS.commitDelay] = commitDelay
 
     Object.freeze(this)
   }
@@ -96,12 +121,27 @@ export default class DatabaseMigrator {
    *         descriptors.
    */
   executeMigration() {
+    let keepAliveObjectStore = this[FIELDS.database].
+        createObjectStore(this[FIELDS.keepAliveObjectStore])
+    let keepAlive = new KeepAlive(() => {
+      return keepAliveObjectStore
+    }, this[FIELDS.commitDelay])
+    
+    // start the keep-alive to handle the delay introduced by promises when no
+    // records are fetched before schema upgrade
+    keepAlive.requestMonitor.monitor(keepAliveObjectStore.get(0))
+    
     return migrateDatabase(
       this[FIELDS.database],
       this[FIELDS.transaction],
       this[FIELDS.schemaDescriptors],
-      this[FIELDS.currentVersion]
-    )
+      this[FIELDS.currentVersion],
+      keepAlive,
+      this[FIELDS.keepAliveObjectStore]
+    ).then(() => {
+      keepAlive.terminate()
+      this[FIELDS.database].deleteObjectStore(keepAliveObjectStore.name)
+    })
   }
 }
 
@@ -116,12 +156,16 @@ export default class DatabaseMigrator {
  *        descriptors of the database schemas for various versions, sorted in
  *        ascending order by the version number.
  * @param {number} currentVersion The current version of the database schema.
+ * @param {KeepAlive} keepAlive Utility keeping the transaction alive while the
+ *        promise callbacks are executing.
+ * @param {string} keepAliveObjectStore The name of the object store used to
+ *        create keep-alive requests.
  * @return {Promise<undefined>} A promise that resolves when the schema is
  *         upgraded to the greatest version specified in the schema
  *         descriptors.
  */
 function migrateDatabase(database, transaction, schemaDescriptors,
-    currentVersion) {
+    currentVersion, keepAlive, keepAliveObjectStore) {
   let descriptorsToProcess = schemaDescriptors.filter((descriptor) => {
     return descriptor.version > currentVersion
   })
@@ -133,13 +177,17 @@ function migrateDatabase(database, transaction, schemaDescriptors,
   return migrateDatabaseVersion(
     database,
     transaction,
-    descriptorsToProcess[0]
+    descriptorsToProcess[0],
+    keepAlive,
+    keepAliveObjectStore
   ).then(() => {
     return migrateDatabase(
       database,
       transaction,
       descriptorsToProcess,
-      descriptorsToProcess[0].version
+      descriptorsToProcess[0].version,
+      keepAlive,
+      keepAliveObjectStore
     )
   })
 }
@@ -153,18 +201,28 @@ function migrateDatabase(database, transaction, schemaDescriptors,
  *        transaction.
  * @param {(DatabaseSchema|UpgradedDatabaseSchema)} descriptor Schema
  *        descriptor of the version to which the database is to be upgraded.
+ * @param {KeepAlive} keepAlive Utility keeping the transaction alive while the
+ *        promise callbacks are executing.
+ * @param {string} keepAliveObjectStore The name of the object store used to
+ *        create keep-alive requests.
  * @return {Promise<undefined>} A promise that resolves once the database has
  *         been upgraded to the schema described by the provided schema
  *         descriptor.
  */
-function migrateDatabaseVersion(database, nativeTransaction, descriptor) {
+function migrateDatabaseVersion(database, nativeTransaction, descriptor,
+    keepAlive, keepAliveObjectStore) {
   let transaction = new Transaction(nativeTransaction, () => {
     return nativeTransaction
-  })
+  }, keepAlive)
   let objectStores = descriptor.fetchBefore || []
   
   return fetchRecords(transaction, objectStores).then((recordsMap) => {
-    upgradeSchema(database, nativeTransaction, descriptor)
+    upgradeSchema(
+      database,
+      nativeTransaction,
+      descriptor,
+      keepAliveObjectStore
+    )
     
     if (descriptor.after) {
       return Promise.resolve(descriptor.after(transaction, recordsMap))
@@ -181,15 +239,20 @@ function migrateDatabaseVersion(database, nativeTransaction, descriptor) {
  *        transaction.
  * @param ((DatabaseSchema|UpgradedDatabaseSchema)) descriptor Schema
  *        descriptor of the version to which the database is to be upgraded.
+ * @param {string} keepAliveObjectStore The name of the object store used to
+ *        create keep-alive requests.
  */
-function upgradeSchema(database, nativeTransaction, descriptor) {
+function upgradeSchema(database, nativeTransaction, descriptor,
+    keepAliveObjectStore) {
   let objectStoreNames = Array.from(database.objectStoreNames)
   let newObjectStoreNames = descriptor.objectStores.map((objectStore) => {
     return objectStore.name
   })
   objectStoreNames.forEach((objectStoreName) => {
     if (newObjectStoreNames.indexOf(objectStoreName) === -1) {
-      database.deleteObjectStore(objectStoreName)
+      if (objectStoreName !== keepAliveObjectStore) {
+        database.deleteObjectStore(objectStoreName)
+      }
     }
   })
 
@@ -334,6 +397,32 @@ function normalizeFetchBeforeObjectStores(objectStores) {
       return objectStore
     }
   })
+}
+
+/**
+ * Generates a non-conflicting object store name to use to keep the transaction
+ * alive as long as needed.
+ * 
+ * @param {((DatabaseSchema|UpgradedDatabaseSchema)[]|Object[])}
+ *        schemaDescriptors The database schemas for database versions to
+ *        validate, sorted by version number in the ascending order.
+ * @return {string} The name of the object store to use during the
+ *         {@code versionchange} transaction to keep the transaction alive as
+ *         long as needed.
+ */
+function generateKeepAliveObjectStoreName(schemaDescriptors) {
+  let longestName = ""
+  
+  schemaDescriptors.forEach((versionSchema) => {
+    versionSchema.objectStores.forEach((objectStoreSchema) => {
+      let objectStoreName = objectStoreSchema.name
+      if (objectStoreName.length > longestName.length) {
+        longestName = objectStoreName
+      }
+    })
+  })
+  
+  return `keep-alive ${longestName}`
 }
 
 /**
